requirements.txt
boto3
s3fs
pandas
sqlalchemy
psycopg2-binary
pyarrow
numpy